% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_functions.R
\name{find_cpg_sites}
\alias{find_cpg_sites}
\title{Find CpG Sites for a Gene}
\usage{
find_cpg_sites(gene_name)
}
\arguments{
\item{gene_name}{Name of the gene (e.g. "TP53")}
}
\description{
This function returns CpG sites associated with a gene
}
\examples{
find_cpg_sites("TP53")
}
