% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCPGFinder.R
\docType{package}
\name{GeneCpGFinder-package}
\alias{GeneCpGFinder-package}
\alias{GeneCpGFinder}
\title{GeneCpGFinder: Fast Genomic Information Search for Human Genes}
\description{
A lightweight R package for instant gene information search using local database of 78,656 human genes. Provides ultra-fast access to genomic coordinates without internet connection.

A lightweight R package for instant gene information search
using local database of 78,656 human genes.
Provides ultra-fast access to genomic coordinates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Islam-DS/GeneCpGFinder}
  \item Report bugs at \url{https://github.com/Islam-DS/GeneCpGFinder/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/Islam-DS/GeneCpGFinder}
  \item Report bugs at \url{https://github.com/Islam-DS/GeneCpGFinder/issues}
}

}
\author{
\strong{Maintainer}: Md. Mahbubul Islam Bioinformatics Research Assistant \email{mahbubulislam331@gmail.com}

}
\keyword{internal}
