% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCPGFinder.R
\name{get_cancer_genes}
\alias{get_cancer_genes}
\title{Get Common Cancer Genes}
\usage{
get_cancer_genes()
}
\value{
Character vector of cancer gene symbols
}
\description{
Returns a list of frequently studied cancer genes
}
\examples{
get_cancer_genes()
}
