% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_cancer_database.R
\name{get_real_cancer_cpg_database}
\alias{get_real_cancer_cpg_database}
\title{Pre-downloaded Real Cancer Genes CpG Database}
\usage{
get_real_cancer_cpg_database()
}
\value{
Real cancer genes with CpG sites from actual databases
}
\description{
Pre-downloaded Real Cancer Genes CpG Database
}
