% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_cancer_database.R
\name{get_real_cpg_sites}
\alias{get_real_cpg_sites}
\title{Get Real CpG Sites for Gene}
\usage{
get_real_cpg_sites(gene_name)
}
\arguments{
\item{gene_name}{Gene symbol}
}
\value{
Real CpG sites information
}
\description{
Get Real CpG Sites for Gene
}
