% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCPGFinder.R
\name{search_multiple_genes}
\alias{search_multiple_genes}
\title{Search Multiple Genes}
\usage{
search_multiple_genes(gene_list)
}
\arguments{
\item{gene_list}{Vector of gene symbols}
}
\description{
Search for multiple genes at once
}
\examples{
search_multiple_genes(c("TP53", "BRCA1", "EGFR"))
}
